refactor(rich-ui): modularize RichUI controller into layered architecture
Phase 2: Move monkey patches from ruby_rich_patches.rb into individual
extensions/ modules with explicit apply! calls (viewport_selection,
transcript_plain, markdown_table_adapter). Replace instance_variable_get/set
with proper accessors across shell, status_view, sidebar, and controller.

Phase 3: Introduce BaseComponent (shared rendering primitives: muted,
colored, status_marker) to eliminate duplicated ANSI helpers across
sidebar panels and dialogs.

Phase 4: Extract ViewRenderer module with 23 stateless formatting helpers
(tool output, args, markdown, config dialogs, approval parsing) out of
RichUIController — controller shrinks from 1112 to 822 lines (-26%).

Phase 5: Add EntryTracker for id-based tool call tracking, replacing the
fragile @tool_ids stack with semantic API (register_tool, pop_tool_id).

Fix: Override convert_table (not table) in MarkdownTableAdapter to match
Kramdown's TerminalConverter dispatch. Use @width instead of
@options[:width] for terminal width lookup.

Tests: Update slash command spec to stub on_model_switch. Update RichUI
specs to match refactored sidebar panel structure (Work/Tasks/Context).
